refactor: split per-level right-column renderers into their own file
Move the per-level right-column renderers (renderRightResourceTypes
through renderFallbackYAML) out of view_right.go into a new
view_right_levels.go, keeping the shared column plumbing, predicates,
and kind-label helpers in place and bringing view_right.go back under
the 800-line cap.
